+++ tests/Traits/AdminMethodsIndexTest.php
@@ -5,8 +5,13 @@ namespace Larrock\Core\Tests\Traits;
 use DaveJamesMiller\Breadcrumbs\BreadcrumbsServiceProvider;
 use Larrock\ComponentBlocks\BlocksComponent;
 use Larrock\ComponentBlocks\LarrockComponentBlocksServiceProvider;
+use Larrock\ComponentCategory\LarrockComponentCategoryServiceProvider;
+use Larrock\ComponentFeed\FeedComponent;
+use Larrock\ComponentFeed\LarrockComponentFeedServiceProvider;
 use Larrock\Core\LarrockCoreServiceProvider;
 use Larrock\Core\Tests\DatabaseTest\CreateBlocksDatabase;
+use Larrock\Core\Tests\DatabaseTest\CreateCategoryDatabase;
+use Larrock\Core\Tests\DatabaseTest\CreateFeedDatabase;
 use Larrock\Core\Traits\AdminMethodsIndex;
 use Larrock\Core\Traits\ShareMethods;
 use Orchestra\Testbench\TestCase;
@@ -29,6 +34,12 @@ class AdminMethodsIndexTest extends TestCase
 
         $seed = new CreateBlocksDatabase();
         $seed->setUpBlocksDatabase();
+
+        $seed = new CreateCategoryDatabase();
+        $seed->setUpCategoryDatabase();
+
+        $seed = new CreateFeedDatabase();
+        $seed->setUpFeedDatabase();
     }
 
     protected function getPackageProviders($app)
@@ -37,6 +48,8 @@ class AdminMethodsIndexTest extends TestCase
             LarrockCoreServiceProvider::class,
             LarrockComponentBlocksServiceProvider::class,
             BreadcrumbsServiceProvider::class,
+            LarrockComponentFeedServiceProvider::class,
+            LarrockComponentCategoryServiceProvider::class
         ];
     }
 
@@ -44,6 +57,8 @@ class AdminMethodsIndexTest extends TestCase
     {
         return [
             'LarrockBlocks' => 'Larrock\ComponentBlocks\Facades\LarrockBlocks',
+            'LarrockFeed' => 'Larrock\ComponentFeed\Facades\LarrockFeed',
+            'LarrockCategory' => 'Larrock\ComponentCategory\Facades\LarrockCategory',
             'Breadcrumbs' => 'DaveJamesMiller\Breadcrumbs\Facades\Breadcrumbs'
         ];
     }
@@ -57,6 +72,15 @@ class AdminMethodsIndexTest extends TestCase
     public function testIndex()
     {
         $test = new AdminMethodsIndexMock();
+        $this->assertEquals('larrock::admin.admin-builder.categories', $test->index()->getName());
+        $this->assertEquals(1, $test->index()->getData()['categories']->total());
+
+        $test = new AdminMethodsIndexMockBlock();
+        $this->assertEquals('larrock::admin.admin-builder.index', $test->index()->getName());
+        $this->assertEquals(1, $test->index()->getData()['data']->total());
+
+        $test = new AdminMethodsIndexMockBlockNotPosition();
+        $this->assertEquals('larrock::admin.admin-builder.index', $test->index()->getName());
         $this->assertEquals(1, $test->index()->getData()['data']->total());
     }
 }
@@ -69,6 +93,31 @@ class AdminMethodsIndexMock
 
     public function __construct()
     {
+        $this->config = new FeedComponent();
+    }
+}
+
+class AdminMethodsIndexMockBlock
+{
+    use AdminMethodsIndex, ShareMethods;
+
+    protected $config;
+
+    public function __construct()
+    {
         $this->config = new BlocksComponent();
+    }
+}
+
+class AdminMethodsIndexMockBlockNotPosition
+{
+    use AdminMethodsIndex, ShareMethods;
+
+    protected $config;
+
+    public function __construct()
+    {
+        $this->config = new BlocksComponent();
+        $this->config->removeRow('position');
     }
 }